Extract registration methods in dedicated controller
The UsersController was becoming too big and thus, harder to maintain.
Since I would probably have to add more and more methods to it, I prefer
to split this controller.

I also renamed the `registration()` method because it didn't convey the
same meaning in this new controller.
