+++ src/views/registrations/new.phtml
